DID A TON OF WORK TODAY!!!
I labeled so many commits so far that when I am fixing my code I think oh was that a architectural technical debt. The same goes for when I am writing my commits `What would I label this commit Strong-ATD, Weak-ATD or Non-ATD?`. Normally the answer  is non-ATD but this time it is actually strong-ATD because I realized I did a very big mistake on the backend which was fixed though so all good now!!!